Move ManualTriggerEntity
